==> 5-fastText-features.ipynb <==
from fastText import train_supervised, train_unsupervised
import itertools
import json

def load_data(fname):
    data = []
    with open(fname) as f:
        return [json.loads(l.strip()) for l in f]

train = load_data('train_for_vectors.jsonl')
test = load_data('test_for_vectors.jsonl')

# fit unsupervised model on data, starting from pretrained vectors
# need to do this since there is no option to read in .vec files
# vectors from https://fasttext.cc/docs/en/english-vectors.html

unsup_path = 'unsup.train'
with open(unsup_path, 'w') as fh:
    for d in itertools.chain(train, test):
        fh.write(f'{d["text"]}\n')
        
model = train_unsupervised(unsup_path, pretrainedVectors='crawl-300d-2M.vec.zip')

def write_vectors(model, data, path):
    with open(path, 'w') as fh:
        for d in data:
            text = d['text'].replace('\n', ' ')
            d['vector'] = list(float(v) for v in model.get_sentence_vector(text))
            fh.write(f'{json.dumps(d)}\n')

write_vectors(model, train, 'train_with_vectors.jsonl')
write_vectors(model, test, 'test_with_vectors.jsonl')
